feat: autofight for wait spells
The need to specifically hit `.` is a bit of an interface screw more
than anything; this commit enables just triggering autofight (`p` etc)
if the wait spell is quivered to continue casting it. This has been on
my list for a while but now that there are two of them it became more
important.

As part of this commit, I also made it clearer in the targeter that
casting it that way would restart the spell. Possibly should be included
in `z` as well for people who have the Maxwell's targeter disabled.