Remove the chance of paralysis when ending berserk
Paralysis was a rare occurence on ending berserk, and under Trog was
particularly unlikely to happen except very early on, due to the
piety-based chance to prevent it. It was slightly more relevant for
non-Trog berserk, but since that's much less reliably repeatable (and
weaker due to not being extended on kill), remove the chance of paralysis
to simplify the effect overall.