Improve some unrand descriptions
Add DESCRIP lines for the Wrath of Trog, necklace of Bloodlust, ring of
Shadows and amulet of the Air. Unrands generally could probably use a full
pass to put pure mechanics information in DESCRIP lines, free up the
actual unrand description portions to be primarily flavourful, and possibly
to use the new ARTP syntax of showing the inscription alongside the
description (such as "SInv: It lets you see invisible"). Also fixes a typo
in the cloak of the Thief description (via geekosaur, thanks!).