Wereblood: only heal on adj kills
A duration spell has to be balanced very carefully, to encourage it
to be used situationally, rather than on all the time (since spells
in DCSS do not work well for 'always-on' effects). Adding healing
to Wereblood was intended to make it more useful in the early game,
so that characters would have the 'sustain' needed to keep fighting
multiple enemies and to make use of the slaying, but it seems to
have made it too generally applicable. Let's limit it a bit.