fix: move newchar quiver setup later (dang3rdave)
It turns out that _setup_generic happens before the player's init file
is read for real, and so anything that is waiting on early setup calls
or lua isn't reliable at this point in character creation. This includes
the new fire order options, because there's an overly elaborate spell
name cache thing that isn't in place for the first rc read (and some
maybe-related but not fully identified issues for abilities). To solve
this I have moved new character quiver setup later in the sequence. It
might eventually be better to simplify spell/ability startup so it's
more reliable.