Fix a crash when Djinn learn too many spells
In rare cases, Djinn Wanderers could start off with 9 or 10 spells, which
would cause them to hit the cap of known spells at level 25 or 27, and
crash when trying to learn their next spell. To rescue any games trapped in
this situation, make them forget spells from the start of their innate
set to make room for the later ones. A later commit should adjust Wanderer
starts to prevent Djinn from giving out so many spells and causing the
issue in the first place.