feat: Spell background adjustments part I (kate-, hellmonk)
Book backgrounds have for a long time presented friction with the notion
that a background should be a starting kit, and not completely determine
a character's arc. With many books containing L5 spells many could carry
comfortably to the midgame. In part this was necessary to make them fun:
with spell books being large and somewhat rare, finding "the right book"
could take some time and without the level of progression found in
current books they would be sunk cost traps.

The recent changes to book items smooth out spell drops, giving more
opportunity for characters to adapt. This opens up space to make spell
backgrounds more of a starting kit and less of a complete package. Some
of the backgrounds would be awkward with a blunt delete all L5 spells
from spell starts, so this commit starts the process incrementally. This
is also a good opportunity to reduce the number of schools in each start
and the total number of spells even if there are no L5 spells, also in
support of this goal.

Ne, FE, and Wr are untouched, as chopping off their L5 spells leave the
books a bit too thin. They will be visited in future commits (possibly
in 0.28).

AE loses lightning bolt. There's a bit of an aesthetic loss in that
lbolt is quite iconic, but the start still has a strong collection of
spells that can support developing in various directions based on the
dungeon. Indeed, there are many players who already ignore lbolt in the
book.

AM loses Leda's. The AM book had four (4) schools for a background that
notionally also wants to train a weapon. At level 4 dual school Leda's
was the hardest to reach of these spells, so its absence won't be too
felt.

EE loses LRD. People kill OOFs with an acid wand and LRD, come on.
Without it the "jank stabber", "pick up some kind of tmut either form or
otherwise", and "be a blaster" possibilities invited by the book have
more equal appeal.

IE loses Metabolic Englaciation. The loss makes repeated Hailstorm
harder to set-up from the book alone, but it remains very strong.

Su loses Summon Ice Beast. Still very strong but reduces the book to a
three school book.

VM loses Ignite Poison. This reduces the number of schools in the book
to 4 (still high) and returns VM to an era of needing to branch out
sooner, especially to deal with resistant threats. This is much like
VM's past, but with the new opportunities provided by more frequent
spell drops it should be less of a "VM sucks" vibe.