feat: Add a vault utility function
The lua function dgn.random_entry_arg() can be used to add an argument
to each entry in a delimited string. This can be useful when adding
weights or terms to strings defined for monster and item definition
purposes so that said strings can be re-used without repeating their
contents in multiple places. This will be used in a subsequent commit
that updates the Ice Cave maps.

Also improve and reformat comments for other functions in this file.