feat: Have all Ice Cave maps use the same weight
Maps currently have a weight of either 5 or 6, making them individual
place at nearly the same frequency. The weight of 6 was used for the
five "small cave" maps to make them place twice as often as the three
"sprawling cave" maps that use weight 5. This isn't necessary, however,
since the themes of these maps are strongly overlapping. The two map
sets share similar monsters, have similar difficulty as a whole, and
nearly all of the "small" maps are in fact quite big. This commit has
all maps use the default weight of 10.