fix: prompt when summoning would be a bad plan (12602)
Previously we warned the player about summoning with OTR up, but we
should really check for bog and vortex too. If this results in a
profusion of prompts rude_stop_summoning prompt can be changed to take
an optional pal representative and made a bit smarter, but I expect
that this change won't generate very many more prompts, given the
non-overlap of spell schools.