fix: process ring unequips when sacrificing hand (12587)
This adds a bit of a special case to avoid charging the player a removal
penalty if they keep the ring on because of the sacrifice swap
hand/tentacle. A simpler solution would be to always impose the swap
penalty, but this would come at the cost of adding more prompts to sac
hand for special cases. This is cleaner and bothers the player less.