New wizlab: Maxwell's Workshop
Features constructs and big beefy guys with great maces. Boss is an electric
golem. Loot includes several items with bad or dangerous properties. This vault
still needs some serious beautification work as well as some balance tweaking,
and the construct set in doroklohe needs to be removed, which will be done in
future commits.