feat: add a game menu
This commit introduces a serviceable in-game menu of the sort you might
expect in games that aren't 20 year old open source roguelikes. It's
possible that some people will hate this nod to modernism, but I think
it's actually kind of useful, and helpful to new players. Will take
suggestions on what should or shouldn't be here. This also may be a
useful place to get to a hypothetical future in-games options editor.

Unfortunately, I don't think it's a great idea to put this on ESC on the
default keybinding, since that has for so long been a noop when at the
main game UI. So for now it replaces macro editing on `~` (and hitting
`~~` brings you to the macro editor).

For tiles, this commit reuses existing tiles (partly drawing on ones for the
local tiles system toolbar, which has some functionality overlap with this
menu). Some of these might benefit from artist attention.