feat: try to map more currency symbols to CMD_LIST_GOLD
I am not sure whether this will (a) build on every platform, or (b) work
on every build target. When testing on a US keyboard layout on mac, I
can consistently get these to work on console, but not on the other two.
However, I think this is because I need to enter them as complex chorded
keystrokes involving number keys, where local tiles number key
processing interferes. Not sure what's happening in webtiles at all, but
on a US layout these definitely don't work for me. We'll see what, if
anything, happens on other layouts.