fix: avoid a crash on upgrading randbooks after ca9b790695f0
Randbooks don't have artefact props, and ca9b790695f0 added some
explicit checks for this. Most code is careful about this, but there
were a bunch of fairly old save compat checks that were not. It probably
would be fine to loosen up the artefact_property check, but it's also
easy to prevent the call for the save compat cases. I've made the
artefact_property check more explicit as well, so that it might be
easier to guess what's going on here.