Adjust randomisation in some kennysheep vaults
ks_random_setup is used in a number of vaults to heavily randomise various
features, but the randomisation of water in particular could have some
extremely strange results in certain vaults, with large pools being
replaced with huge numbers of statues. Instead only allow water to be
substituted for floor, lava, or trees (this also has the bonus of making
the behaviour match how the function's comment claims it works).