Rework a transporter vault
The vault gammafunk_worm_tunnel is a memorable challenge, but it's
conspicuously large for a challenge vault of this type. This commit
drastically reduces its size, slightly changing layout and adjusting
swamp worm placement slightly so it remains a significant challenge.

It also adapts the vault to place in earlier dungeon and Lair, using net
traps instead of Zot traps with an additional miasma cloud generator
before D:14. The net traps are far less lethal than Zot traps, but still
trap players, give the worms more opportunities to slurp. The additional
miasma generator near the exit chamber door is used when when net traps
are used to slow down a player using a blink scroll a bit more than a
net trap at that location would. When Zot traps are used, a player doing
this must run the risk of blinking onto a Zot trap, so a trap there is
appropriate.

Loot and additional worm placement is revised to scale better with
placement depth. We use trees in Swamp so that this vault better fits in
with the surrounding layout. I opted against using trees elsewhere to
minimize the use of fire magic to break into the vault. In Swamp we're a
bit happier to let fire magic take over the place.