Simplify a lua test and reindent
I'm not sure if this test needs to exist as a standalone test given our
existing map connectivity checks, but so long as it does exist, it needs
to properly ignore any shoals encompass maps for the branch end. A
previous fix to the dlua API allows the existing code to work. This
commit just simplifies its logic and reindents.