Rework the master elementalist vault
The vault mu_elemental_laboratory features a vault-defined "master
elementalist" monster that's themed as having mastered the four forms of
elemental magic. However this monster doesn't pose a very good threat,
certainly not one worthy of a having a vault-defined monster. The vault
itself is not especially challenging either. This commit reworks the
monster and the associated vault to pose a greater challenge and give
greater rewards, reworking it into a rare runed door vault.

The master elementalist gets an HD bump to 18 and boasts a formidable
array of spells: Ozocubu's Refrigeration, Lehudib's Crystal Spear, Chain
Lightning, and Fire Storm. It still has Haste, but loses Blink for
Repell Missiles, since the former doesn't work great for its terrain in
this vault. The monster attempts to generate with the Elemental Staff,
thus giving him formidable elemental resistances. If the Elemental Staff
is unavailable, it instead gets a staff of air, and its robe gains the
resistance ego, thus it will always have at least rF, rC, and rElec. The
master elementalist is a formidable spellcaster, but it doesn't have
especially good HP or other defenses, so players with good resistances
who act quickly can take it on.

The rest of the vault monsters have been given a rework to pose a
serious threat for both Elf and Depths, and the elemental-themed terrain
has been enhanced. Loot includes randbook for one of each of the four
elemental schools, as well an artefact jewellery, an enhancer staff, and
a manual, all theme from one of the elemental schools.

The master elementalist monster definition is moved to a function, since
it will be re-used for a forthcoming guarded unrand vault.