vaults: Adjust the despoilered loot in portals
For the most part, change the loot lists to use standard dgn.* constants
instead of identifying everything. The identification in early portals
would eat up a large portion of the remaining item ID game, which would
be a shame. Usingn the dgn.* constants does mean that portal loot has a
different distribution than the floor, but it's not a vault by vault
absurd spoiler.

Closes #1855