Don't let Folly's brilliance stack with the potion
To make its behaviour slightly clearer, since as originally implemented
the MP reduction didn't stack but the enhancer did. Its permabrilliance
is now technically separate from regular Archmagi, so would stack with it
if a non-body-armour source of Archmagi existed. Potions of brilliance
are marked as temporarily useless with Folly equipped.