feat: tweak the abyssal woods Lair ending
For hellmonk_lair_end_abyssal_woods, slightly adjust the monster set,
introduce chances to have shortcut entrances, tweak the use of demonic
trees, add some use of floor tiles, and clean up DES syntax.

For the monsters, introduce a 5th monster tier below the "boss" tier on
glyph 6. Move the small chance of large abominations formerly on 6 to
this new 5 glyph, which it now shares with ugly things. The latter are
appropriate to the map, and although they are common enemies later, we
don't overuse them here. Then 6 remains dedicated to thrashing horrors,
whose numbers we constrain somewhat. Now only 2-4 of those can now
place, with an average of 3, since we only place these in the final
chamber, and we place a few 5 glyph monsters throughout the map.

For layout, introduce two potential shortcuts, each with an independent
33% chance, one that leads something part the starting area and another
on the far side that leads in close to the final chamber. Instead of a
forest of all demonic trees, use mostly regular trees, gradually
introducing demonic trees at an increasing rate until the final chamber,
which uses all demonic  trees. Also add Abyss-themed floor tiles in this
final chamber. Adjust a few loot locations due to the new shortcuts.
Closes #1967.