feat: Rework parts of a Lair ending
For gammafunk_lair_ancient_temple, rework the way the temple is
"ruined", tweak monster placement and loot, and simplify syntax.

For layout, we greatly simplify the crumbled wall sections. Each of the
three walls now has an independent 50% of being "collapsed". The
collapse wall sections avoid clumping the "broken" wall pieces together
and add dry land to facilitate their usage by players.  We add an inner
chamber with possible door to house the loot and provide more structure.
Introduce very light ruining of the fixed wall portions and chamber to
prove a few small entrances into the temple. Finally, we introduce
random dry land throughout the map to give players better footing while
fighting closer to and inside the temple.

For monsters, we reduce the average monster count by about 33%. The map
was not extreme in terms of monster placement, but it's been the Lair
end with the most kills by a wide margin for a long time, in large part
due to how effective the merfolk avatar is at punishing incautious
players. We now place a couple more of the Shoals themed threats on
average, since those were somewhat underused. Add snapping turtles to
the list of Shoals monsters, which will help prevent an overall increase
in wind drakes; those seem to pose the biggest challenge of the
non-avatar monsters in the Shoals set. For the Lair natives, add blink
frogs and steam dragons, and remove the very trivial water moccasins.

For loot, we replace some of the normal items with thematic items that
include 2-4 good loot potions and a single water/Shoals themed item. The
water/Shoals themed item includes throwing nets, trident weapons, and
phial of floods, and small chances for the three unrands Captain's
Cutlass, Trident of the Octopus King, and a ring of the Octopus King.