feat: Adjust the due_jungle_book Lair end vault
We don't have a lot of data for the new version of this map, and its
newness means players are more likely to make mistakes, but early
indications are that people die to it in practice a lot. It does feature
a somewhat higher monster count as well as some legitimately nasty
monsters not commonly seen in Lair. This commit reduces the monster
count by about 20% and fixes up portions of the layout.

For monsters, we reduce the bands spawning in the major combat areas
from 6 monster total on-average to 3. That way the focus is on the final
necromancers in the two end chambers. We also reduce the random spawns
throughout the map, concentrating them on the earlier entryway area,
since none of the bands spawn there. That way the map is less likely to
be empty until one of the bands are encountered.

For layout, we remove the many additional entry connectors, since just
one is necessary. The additional ones likely increased the number of
vetoes. This might increase the chance that the vault is located at the
end of a passage, but this is fitting for the "spooky jungle castle"
theme the map now has.