Add new entry subvaults for the Hall of Blades
Committer's Note: I'm not adding nicolae_blade_animated_armoury, since
animated armour monsters designed to be a player ally, not a compelling
hostile monster. Melee-only slow monsters are something we've almost
completely removed at this point. There's not an obvious design for
hostile animated armour monsters that's sufficiently unique, but if that
changes, a vault like this can be reevaluated.