feat: macro menu, the final form
4a6b093f0 filled a big need in macro editing, by allowing macros to be
chosen from a menu, and viewed all at once. However, it was pretty
incomplete in a number of ways, keeping the old message-window pseudo-ui
for actual editing. This had a number of problems.  97433f2 exaggerated
some of the problems of this approach -- since the menu is just used to
pick which macro to edit, there's a lot of keystrokes to restart the
process if you want to edit more than one at once, or select the wrong
one, etc.

This commit attempts to finally provide a coherent macro editing UI.
* Editing individual mappings is now done via popus from the main macro
  edit menu. All of the message-window code has been converted to
  Menu-based code. Text entry uses a generalized version of the
  title_prompt input box previously only used for spell search.
  Hopefully most aspects of this will be much more convenient than the
  old approach.
* To make it easier to use this menu with arrow keys, several
  meta-functions have also been Menu-ified.
* Deleting individual mappings is possible from the main list.
* Macros vs the various keymap modes are now handled within the menu;
  getting at keymap editing is a bit hidden to keep people from
  preferring those over macros. (Basically every "this key doesn't work"
  report I've run across was someone using a keymap instead of a macro.)
* key hints throughout these menus have been (I hope) improved from the
  message window version.
* This UI has a proper help screen.

Main TODO that I am aware of: it would be nice to show a cursor when
editing in local tiles / console. (Webtiles is fine.)

Some subsidiary stuff:
* The `Menu::title_prompt` API has been massively improved so that it
  works in webtiles in a general way, and setting a custom prompt always
  works.
* min Menu width now (sort of) works in webtiles
* various minor tweaks / bugfixes to menu code (including some weird
  stuff with spell search)