fix: streamline new action entry
Treat new actions started by `~` and by key in the same way, jumping
right into editing. To get raw entry, you can use enter to get out
to the menu with a blank mapping; esc still cancels. (I'm not sure if
this behavior is very intuitive and it is currently somewhat hidden, but
raw key entry may also be a fairly niche feature.)

Doing this required adding an on_show callback to Menu, and tweaking the
title_prompt api to provide a way to detect whether text entry was
cancelled or not.