fix: default allow_extended_colours to true
This *may* still be a terrible idea, but I'm not sure we'll find out
without trying. Some of the historical reasons for the old defaults are
no longer an issue, like old mac osx terminals. One remaining one is
that this may decrease the default readability of ttyrecs. We'll see...