fix: Place a Swamp exit in bugged saves (Sergey)
The Swamp exits destroyed by eleionoma due to the issue fixed by 0d5cf04
need to be restored in affected saves, otherwise those games always
crash on Swamp:1. This commit adds a save compat function _ensure_exit()
that's analogous to the _ensure_entry() function used for ensuring
branch entries and calls it for Swamp:1. We find whether Swamp is
missing its exit, and if so, find a floor or shallow water square that's
as close as possible to the first down stairs, thus helping to ensure
that our exit is placed at a connected location.