fix: limit `passable` tag masking to Vaults
This tag is a bug magnet outside of Vaults, because it makes it very
easy to trick the builder into generating disconnected zones. In some
vaults these issues had been fixed by explicitly conditionalizing the
tag, but there's not a lot of use-cases outside of Vaults layouts that
can't be better handled by other techniques, so simplify things on the
vault designer here. I've also updated the docs to more clearly warn
designers about the pitfalls.