fix: simplify macro action entry (Yermak)
After this commit, entering an action while doing a macro binding
returns to the main macro menu (or exits in the case of quick-add),
while esc gets you to the mapping menu. This reduces the keystrokes used
when setting at the slight cost of keystrokes when cancelling.

Two remaining issues: (i) when editing an old macro, it first brings you
to the mapping submenu. This could go right into editing the action, at
the cost of making the submenu c/R options harder to get to. (ii) What
should this UI do on enter with a blank action? Right now this cancels,
but I'm not sure this is very intuitive. Other possibilities include
having this case clear the macro, and disabling enter with an empty
action.

A few other minor tweaks to display stuff.