fix: quick macro keybinding add, take 3
Unfortunately, ctrl-m is equivalent to return, and so this turns one of
the standard noops (that people use to clear mores) into a menu key by
accident. This third attempt moves the travel speed toggle to ctrl-t,
and puts quick add on ctrl-e. I *think* this should be fine in webtiles.
(Unfortunately, it does mean that windows webtiles users will need to
use *t for travel speed toggle, because it seems that most windows
browsers eat ctrl-t in a way that isn't easy to override.)