fix: Adjust three transporter vaults again (advil)
A recent report I'd seen for the vault emtedronai_trogs_sanctum showed
it generating areas only reachable via its transporters, effectively
sealing off access to its entrance transporter. This report lead me to
make 97d24401, which put borders of floor around this vault and two
others, since it seemed there was no other way to prevent these
problems. However advil informs me that the problem with
emtedronai_trogs_sanctum was due to it not conditionally applying its
`passable` tag only for the Vaults branch. The `passable` tag was added
for the Vaults layout to allow features like transporters place in
Vaults rooms, but when enabled in other branches it leads to the layout
generator to incorrectly think that areas inside the transporter vaults
are connected to the vault entries without using a transporter. Hence
2170ac5a makes `passable` now only apply in the Vaults branch.

advil also informs me that now that we can't misapply `passable`, these
transporter vaults will both never generate such disconnected areas nor
will passage through the transporter ever be required to reach the rest
of the level, specifically because the internals of these vaults are
marked with the `opaque` tag as of 97d24401. Hence this commit restores
the behaviour before that commit which only applied the border of floor
around these vaults in the Vaults branch; it's still necessary in that
branch for them to place properly as rooms. Now the vaults will better
integrate into non-Vaults levels in that they can embed properly in
surrounding walls yet have their two entries properly connected to the
rest of the level.

Additionally, I've changed the wall types of these vaults to rock when
placing in Vaults, so the layout code of that branch can correctly
substitute the wall to the surrounding wall type of the level.
Interestingly, this seems to allow gammafunk_its_a_trap to properly
place in Vaults. In my previous testing, it always failed to place as a
vaults room even though the other two vaults were able to.

Finally, correctly apply the `transparent` tag to
emtedronai_trogs_sanctum and fix its opaque mask to correctly cover all
internal vault squares. The slightly incorrect mask was added 97d24401,
but since the vault wasn't actually given `transparent`, this error
wasn't seen.

This commit partially reverts 97d24401.