Give Enchanters invisibility again
Corona is not a very fun spell to use. The player is being asked to
spend their time flipping a coin, and, unlike Ensorcelled Hibernation
or Petrify or similar spells, they don't get a satisfying reward.
Instead, they get a roughly 20% accuracy increase, which is really
quite hard to notice. Most of the time, players are better off throwing
stones or doing almost anything else.

So, we'll remove Corona from the Enchanter starting book (that is, leaving
them without a level 1 spell) and give them a single potion of invisiblity
instead. This should be a net buff to the background overall. (It's a nerf
to muen, but mummies should be used to suffering by now.)

Arkane markspersons will be addressed in some way once ranged combat is
reworked. They aren't changed in this commit.

Does not revert 64f0e6ef8afa5d38b99351a6346e47b553adcc76 .