fix: add a debug tripwire for long delays
There have been some cases of delays going on much longer than they
should lately (#1757, #2037 among others) and while a few bugs have been
fixed, the worst cases have resisted debugging or replication. This
commit is an attempt to get more information, while at the same time
protecting the player a bit better. If a rest delay exceeds 500 turns,
this will trigger a crash, and if an explore delay exceeds 2k turns (1/3
of the per-level zot clock) it will also crash. These are intended to be
numbers that can't be triggered on demand ever, so if I've missed
something that allows this, this code should be further amended. The
rest one at least would have been triggered by the case in #2037.