feat: Adjust monster-placing Elf vaults
Most changes are to refactor syntax of recently merged Elf branch
monster vaults by nicolae. We use numeric glyphs for monsters and
simplify statements. A couple vaults have more significant changes, as
described here:

nicolae_elf_defective_weapons_storage:

Instead of having various helpless and inneffective dancing weapons,
rework this vault to place dancing weapons that all have some kind of
permanent debuff (and hence are still "defective"), but have other
aspects allowing them to still be significant threats. Break weapons up
into three tiers by type of ego we'll give as well as type of debuff:
fast acting chaos weapons get blindness, slower acting distortion
weapons get slowness, so they'll be slower than normal speed, and we use
highest quality base types for venom, but make them "damaged" (-1 to -4
enchant) and permanently corroded.

The vault is now a runed door vault since it's fairly dangerous and
having these weird weapons leak out into the level would ruin the theme.
Not to mention that the blind weapons need to stay in a fairly confined
area to be effective. It's layout has gotten a slight trim to completely
reveal the loot inside. Since I worked with nicolae to re-imagine this
vault, we agreed to rename it to
nicolae_gammafunk_elf_defective_weapons_storage.

nicolae_elf_double_double_double_trouble:

This is a runed door vault, but the current layout means a player might
be forced to open the door to reach otherwise disconnected areas in Elf.
We like to avoid forcing players to take runed doors, and although two
blade masters with two double swords is not an especially dangerous
threat that would normally warrant a runed door, it's helpful for the
specific theme of this vault. So mark the doors and sealed areas of the
vault as opaque for connectivity purposes. Add more runed door to keep
the two blademasters separated, increasing the visual theme. Upgrade one
of the blademasters' double swords to either good_item or randart, thus
allowing them to pose a bit more of a threat. Finally, introduce a
star_item and two normal item loot pieces on both sides of the vault,
giving some incentive to open the vault for those who wouldn't benefit
from the weapons.