feat: consumables panel enhancements (clicking, information)
(This is not super well-tested, so I'm not really sure if it is safe to
merge just yet.)

This commit adds the following features to nlavsky's original
consumables panel patch:

* click to show/hide; the leftmost icon when visible is always a hide
  icon.
* right click on the hide icon (only) to get the settings popup
* right click closes the settings popup also. (This should maybe close
* if you click out of this ui, but I wasn't quite sure how to do that.)
* tooltip on consumables gives the item name and inventory letter
* right click on consumables (when in mouse command mode) opens the
  describe popup
* left click (in mouse command mode) just directly triggers the item.
  This is implemented via the actions interface, so doesn't need to add
  much code (as it rests on 6e4aca117). Does this make misclicks too
  easy?

I'm undecided whether this should be minimized or shown by default. It
would be really nice if minimization state at least were stored either
as a cookie or in the save between sessions.

Motivation: The patch as it stands is feature-complete relative to the
original goals (including ones I encouraged), but after testing I
unfortunately don't think it will feel very complete to end users. The
UI has the form of something that should be clickable, so players will
expect it to be clickable. (The tiles even look kind of like local tiles
menus.) Also, players not familiar with the icons will have a lot of
trouble with this: even as a long-time console player I don't know the
icons well. Hence the tooltip.