feat: Allow spectral ego on certain weapons
Currently the spectral ego is only allowed on two-handed weapons or the
one-handed club base type, even for artefacts and vault-defined weapons
that request the spectral ego. Yet the scroll of brand weapon bypasses
this exception, allowing spectral as a result on any one-handed melee
weapon, to help make that scroll more exciting. We already have the
weapon property data to limit spectral to only club or staves on mundane
generated items. This commit removes the handedness restriction for
spectral, thus allowing it to generate on all artefact melee weapons and
on vault-defined melee weapons that specifically ask for that ego. This
won't increase the amount of spectral too much and makes the ego's logic
consistent. It's now simply a melee-only ego that's allowed on
artefacts, like other such egos, but is limited to clubs and staves for
mundane weapon generation.