feat: Adjust some Elf entry vaults
For nicolae_elf_arrival_dilemma, technically we shouldn't require
passage through a runed door like this to the rest of a level, but I'm
allowing it in this case. The right side of the vault is a perfectly
ordinary encounter with four monsters, hence always a reasonable option
for those with basic preparation for Elf. The runed door for this side
fits the "border check" aspect, which allows inexperienced players to
re-evaluate their decision before it's too later. The two choices feel
one-dimensional, and a single high level elf feels a bit sad alone
behind a runed door. So we up the runed door aspect, adding a second
high level elf to the left section and two high quality loot pieces,
also adding one normal quality loot item to the right section. Give this
entry half weight so that it's not too common.

For nicolae_elf_arrival_no_cutting_in_line, ensure we place at least two
monsters in each of the two thematic groups. The odds of rolling no
monsters in the orc group or no monsters in the elf group are not that
small, and if either happens the vault's theme is ruined. This increases
the size of each group by 1.5 monsters on-average, but the groups are
only 3.5 monsters each, and the monsters used here are common for Orc
and late dungeon.

Some glyph and syntax cleanups for this and two other
entry vaults.