fix: avoid static initialization of ability list
This is not necessarily an ideal fix, but is aimed at preventing static
initialization order issues resulting from a7dfd5ec23. After that, the
constructor for the static Options object relies on the static
Abilities_List object, and I suspect this is a source of crashes on some
builds reported in bugs like #2010 (this bug went away, but I have no
confidence that it is fixed).