feat: Clarify bush and briar behaviour
Bushes and briar patches allow beams to pass through them unaffected,
but block explosions and enchantments, which is behaviour that wouldn't
otherwise be obvious to players. Document this aspect of plant physics
in their monster descriptions.

I don't mention the exceptions of beams from ranged Sticky Flame and
Chain Lightning being blocked, as those are niche.