fix: improve some spell data statics
This applies the construct-on-first-use (v2) paradigm to a couple of key
spell data statics, so as to avoid static constructor order issues. This
may improve some hard-to-reproduce issues reported after a7dfd5ec23,
since the Options constructor is called on static initialization but
makes use of this constructor. Spell data isn't in place on the first
call so this may or may not really matter, but it can at least now
reliably be determined that spell data isn't set up yet.