Remove the rune requirement to enter ziggurats
This was added in 07949ec7b9c0 as a replacement for the old gold-based
ziggurat entry cost, rather than as something with a solid justification
in itself. Later discussions (e.g. https://s-z.org/crawl-dev/%23%23crawl-dev-20140929.lg )
suggested that the reason for the 2-rune requirement was as a 'progression'
between Vaults and Zot.

The Zot rune lock exists so that players are forced to battle the difficult
and exciting encounters guarding runes before winning the game, and
the Vaults rune lock likewise exists to encourage players to get at least
one rune earlier, while the earlier S runes are still challenging. The Zig
rune lock, in contrast, has no real justification other than history. So
let's cut it!