fix: don't rest when regeneration is inhibited
The first fruit of 05da648f!

This seems to have come about when 7fb00627a50e removed the vampire
check from player_regenerates_hp, perhaps under the assumption that this
function was not about current regeneration. However, it is used only
for current regeneration state, in order to decide whether resting is
pointless. As of this change, 5 with a monster in sight but out of reach
(e.g. across water) will trigger wait, not rest. Possibly it should also
message, analogous to i_feel_safe's messaging.