fix: default bind F1 to CMD_GAME_MENU
The `~` binding is ok-ish for US keyboards and has some resonance with
the previous binding on this key (which is now a submenu), but it's
pretty non-intuitive, and the symbol typically involves a complex chord
on many international keyboards. At the same time, ESC, which is a
typical binding for this in other games is fairly non-viable in crawl,
because it is used to clear mores and get out of menus very quickly,
while being a noop at the main interface. (The qw bot even has a
function where it exploits this to reliably exit menus.)

This commit adds F1 as a binding for the game menu. This still isn't
perfect, but is a much more likely key for such a thing than `~`. A lot
of people (including myself) like to bind F1 for e.g. casting, and this
is still fine; for advanced users the `~` binding still exists (or it
can be otherwise rebound).

Hopefully this all works across platforms, but I've tested on local
tiles, console, and webtiles and mac.