fix: exclude autofight paths that can't see the target
This is a solution to:
  https://crawl.develz.org/mantis/view.php?id=11905
  https://crawl.develz.org/mantis/view.php?id=12576

In both of these cases, autofight's pathfinding gets into a trap because
under various conditions it prioritizes a move that doesn't keep sight
of the target along a path visible from the starting point. But, because
the search is greedy, it doesn't go back and try an alternative path
when this first attempt fails. Because (for whatever reason) it
prioritizes diagonal moves, the setup usually involves a diagonal, but I
think that switching the prioritization would lead to similar problems
for straight moves. The sequencing here looks very delicate and for the
most part it does seem to make do with a greedy search, so I didn't want
to try messing with this. Rather, this commit introduces a check that
throws out immediate moves where the current target would not be visible
from. I've convinced myself that in theory this is entirely safe to do:
if there's a path at all for autofight, there's a path that keeps sight
of the target at every point. (And of course, potential paths that
don't.) Hopefully this will hold up in practice.