build: try universal builds again
Runners that support cross-compiling are now available, so let's try
that out in CI. It is fairly excruciatingly slow with an empty cache, so
it may end up being necessary to disable this by default just for the
sake of CI not taking forever. But maybe with caching it'll be ok.