Borgnjor wizlab
A split, symmetrical map with ghostly undead on one side and corporeal undead
on the other. Halazid Warlocks at the end serve as the "boss" and are
accompanied by some flying skulls that they might be able to sacrifice.
Additionally, a lost soul sometimes spawns when a monster dies (probably
needs adjustments). Loot is necromancy items and gold.