fix: monsters targeting of invis players (dilly)
If a player goes invisible, monsters that have the player as a foe and
that are sufficiently close to the player will have their throwing and
launcher tracers "fuzzed" to indicate that the tracer hit. This allows
monsters to sometimes shoot at invisible players and works fine if the
player is reachable. But if the player isn't reachable, the logic leads
to monsters performing pointless actions that can lead to monster
collateral kills. For example, if the player goes invisible in front of
a group of hostile monsters behind glass, monsters that can throw or
fire launchers will attempt to shoot at the player, instead hitting
monsters in front of them.

This commit updates the monster logic in two places to prevent this
behaviour. First we don't allow invisible player "hit fuzzing" to work
for monster tracers if the player is not reachable under LOS_NO_TRANS.
Then we require monster throwing/launcher actions to have the same
"friendly monster or hostile monster with a LOS_NO_TRANS reachable foe"
requirement used for spells, most special abilities, potions, and
scrolls. The second change re-uses the result of a check we've already
performed to prevent the tracer from being fired in the first place.
The first change is still good to have, since it's not ever logical for
a tracer to falsely tell a monster that an unreachable player can be
hit.

The logic that governs under what conditions monster take certain
actions could be simplified more than I've done in this commit. Since we
are close to release I want any change to be as minimal as possible, so
this change will only affect throwing/launchers specifically. The other
types of actions either already have the "friendly or near foe"
condition described above, or do their own LOS_NO_TRANS check. I also
suspect that the "friendly" portion of "friendly or near foe" condition
can be dropped, since there's no reason a friendly monster would want to
bypass reachability conditions.

(cherry picked from commit 0adfcbbeba02a92501a471d26d4566247c7335ab)